+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_ar.dart
@@ -1160,6 +1160,11 @@ class DartPdfEditorLocalizationsAr extends DartPdfEditorLocalizations {
   }
 
   @override
+  String sidebarSignatureTrustedViaList(String authority, String list) {
+    return 'Trusted via $authority ($list)';
+  }
+
+  @override
   String get sidebarSignatureUntrustedDetail =>
       'Signer is not from a trusted authority';