+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_fr.dart
@@ -1144,6 +1144,11 @@ class DartPdfEditorLocalizationsFr extends DartPdfEditorLocalizations {
   }
 
   @override
+  String sidebarSignatureTrustedViaList(String authority, String list) {
+    return 'Trusted via $authority ($list)';
+  }
+
+  @override
   String get sidebarSignatureUntrustedDetail =>
       'Signer is not from a trusted authority';