+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_ja.dart
@@ -1134,6 +1134,11 @@ class DartPdfEditorLocalizationsJa extends DartPdfEditorLocalizations {
   }
 
   @override
+  String sidebarSignatureTrustedViaList(String authority, String list) {
+    return 'Trusted via $authority ($list)';
+  }
+
+  @override
   String get sidebarSignatureUntrustedDetail =>
       'Signer is not from a trusted authority';