+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_pt.dart
@@ -1143,6 +1143,11 @@ class DartPdfEditorLocalizationsPt extends DartPdfEditorLocalizations {
   }
 
   @override
+  String sidebarSignatureTrustedViaList(String authority, String list) {
+    return 'Trusted via $authority ($list)';
+  }
+
+  @override
   String get sidebarSignatureUntrustedDetail =>
       'Signer is not from a trusted authority';