+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_ru.dart
@@ -1158,6 +1158,11 @@ class DartPdfEditorLocalizationsRu extends DartPdfEditorLocalizations {
   }
 
   @override
+  String sidebarSignatureTrustedViaList(String authority, String list) {
+    return 'Trusted via $authority ($list)';
+  }
+
+  @override
   String get sidebarSignatureUntrustedDetail =>
       'Signer is not from a trusted authority';