+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_th.dart
@@ -1138,6 +1138,11 @@ class DartPdfEditorLocalizationsTh extends DartPdfEditorLocalizations {
   }
 
   @override
+  String sidebarSignatureTrustedViaList(String authority, String list) {
+    return 'Trusted via $authority ($list)';
+  }
+
+  @override
   String get sidebarSignatureUntrustedDetail =>
       'Signer is not from a trusted authority';