+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_tr.dart
@@ -1140,6 +1140,11 @@ class DartPdfEditorLocalizationsTr extends DartPdfEditorLocalizations {
   }
 
   @override
+  String sidebarSignatureTrustedViaList(String authority, String list) {
+    return 'Trusted via $authority ($list)';
+  }
+
+  @override
   String get sidebarSignatureUntrustedDetail =>
       'Signer is not from a trusted authority';