+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_vi.dart
@@ -1141,6 +1141,11 @@ class DartPdfEditorLocalizationsVi extends DartPdfEditorLocalizations {
   }
 
   @override
+  String sidebarSignatureTrustedViaList(String authority, String list) {
+    return 'Trusted via $authority ($list)';
+  }
+
+  @override
   String get sidebarSignatureUntrustedDetail =>
       'Signer is not from a trusted authority';